+++ R/git.R
@@ -1,0 +1,91 @@
+
+#' Git Repository Helpers
+#'
+#' These functions help use RStudio Connect's configured authorization to query
+#' available branches and subdirectories for deployment using `deploy_repo()`
+#'
+#' - `repo_check_account()` messages whether an account is in use, and then returns that account
+#' - `repo_check_branches()` retrieves which branches are available, returning in a named list
+#' - `repo_check_manifest_dirs()` retrieves which directories contain a `manifest.json`, returning in a named list
+#'
+#' @param client A Connect R6 object
+#' @param host The git repository host (with schema). For example, "https://github.com"
+#' @param repository The git repository to explore or consider deploying
+#' @param branch The git branch to explore for subdirectories
+#'
+#' @rdname git
+#' @name git
+#' @seealso connectapi::deploy_repo
+#' @family content functions
+#' @export
+repo_check_account <- function(client, host) {
+  validate_R6_class(client, "Connect")
+  account <- client$repo_account(host)
+  # TODO: add messaging about which account is being used...
+  if (nchar(account$username) > 0) {
+    message(glue::glue("Accessing '{host}' with account: {account$username}"))
+  } else {
+    message(glue::glue("Accessing '{host}' anonymously."))
+  }
+  return(account)
+}
+
+#' @rdname git
+#' @export
+repo_check_branches <- function(client, repository) {
+  validate_R6_class(client, "Connect")
+  branches_raw <- client$repo_branches(repository)
+  branches_task <- Task$new(connect = client, task = branches_raw)
+
+  task_res <- poll_task(branches_task, callback = NULL)
+  task_data <- task_res$get_data()
+  stopifnot(identical(task_data$type, "git-repo-ref-branch-array"))
+  branches <- purrr::map(task_data$data, ~ .x$branch)
+  purrr::set_names(branches, branches)
+}
+
+#' @rdname git
+#' @export
+repo_check_manifest_dirs <- function(client, repository, branch) {
+  validate_R6_class(client, "Connect")
+  manifest_dirs_raw <- client$repo_manifest_dirs(repo = repository, branch = branch)
+  manifest_dirs_task <- Task$new(connect = client, task = manifest_dirs_raw)
+
+  task_res <- poll_task(manifest_dirs_task, callback = NULL)
+  task_data <- task_res$get_data()
+  stopifnot(identical(task_data$type, "git-repo-branch-manifest-dirs-array"))
+  manifest_dirs <- purrr::map(task_data$data, ~ .x)
+  purrr::set_names(manifest_dirs, manifest_dirs)
+}
+
+#' Deploy a Git Repository
+#'
+#' \lifecycle{experimental} Deploy a git repository directly to RStudio Connect,
+#' using RStudio Connect's "pull-based" "git-polling" method of deployment.
+#'
+#' @param client A Connect R6 object
+#' @param repository The git repository to deploy
+#' @param branch The git branch to deploy
+#' @param subdirectory The subdirectory to deploy (must contain a `manifest.json`)
+#' @param name The "name" / unique identifier for the content. Defaults to a random character string
+#' @param title The "title" of the content
+#' @param ... Additional options for defining / specifying content attributes
+#'
+#' @return A ContentTask object, for use with `poll_task()` (if you want to follow the logs)
+#'
+#' @seealso connectapi::poll_task, connectapi::repo_check_branches, connectapi::repo_check_manifest_dirs
+#'
+#' @family content functions
+#' @export
+deploy_repo <- function(client, repository, branch, subdirectory, name = create_random_name(), title = name, ...) {
+  validate_R6_class(client, "Connect")
+
+  content_metadata <- content_ensure(connect = client, name = name, title = title, ..., .permitted = c("new"))
+
+  deployed_content <- content_item(client, content_metadata$guid)
+  res <- deployed_content$repo_set(repository = repository, branch = branch, subdirectory = subdirectory)
+
+  task <- deployed_content$deploy()
+
+  ContentTask$new(connect = client, content = content_metadata, task = task)
+}