+++ tests/test-integrated.R
@@ -33,7 +33,7 @@ if (nchar(Sys.getenv("CONNECTAPI_INTEGRATED")) > 0) {
     stop("One or both of your integration test servers are not healthy")
   }
 
-  test_dir(rprojroot::find_package_root_file("tests/integrated"), reporter = multi_reporter)
+  test_dir(rprojroot::find_package_root_file("tests/integrated"), reporter = multi_reporter, package="connectapi", load_package = "installed")
 } else {
   message("Not running integrated tests. Set environment variable CONNECTAPI_INTEGRATED=true to run integration tests")
 }